extract rpc request params
