+++ crates/rpc-request-params/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "solders-rpc-requests"
+name = "solders-rpc-request-params"
 version = "0.14.4"
 edition = "2021"
 include = ["/src"]
@@ -7,21 +7,13 @@ license = "MIT"
 repository = "https://github.com/kevinheavey/solders"
 
 [dependencies]
-pyo3 = { workspace = true }
+bincode = { workspace = true }
 serde = { workspace = true }
 serde_with = { workspace = true }
-serde_json = { workspace = true }
 solders-hash = { workspace = true }
-solders-macros = { workspace = true }
-solders-traits-core = { workspace = true }
-solders-traits = { workspace = true }
 solders-commitment-config = { workspace = true }
 solders-rpc-config = { workspace = true }
-solders-rpc-request-params = { workspace = true }
-solders-rpc-request-params-no-config = { workspace = true }
 solders-rpc-version = { workspace = true }
-solders-message = { workspace = true }
-solders-transaction = { workspace = true }
 solders-pubkey = { workspace = true }
 solders-signature = { workspace = true }
-camelpaste = { workspace = true }
+solders-base64-string = { workspace = true }
